[git find-clean-merge] Script to automate merging
Inspired by Rafael Espíndola's merge strategy.

Invoke with:
  git find-clean-merge current-commit commit-to-merge

This script find the most recent revision which merges cleanly, so that
we can get that out of the way, and deal with the smallest conflict we
can by merging the next revision by itself.

It starts searching from "merge-base A B" all the way to B. It
intelligently filters the list of revisions to test by only taking into
account the revisions which change files that have conflicts.